Aktualizacja Funkcji i sprawdzenie ich działania.

+++ src/lab6/P2.java
@@ -1,9 +1,15 @@
 package lab6;
 
+/*
+ * 
+ *  Funkcja Dodawania
+ *   
+ * */
+
 public class P2 {
 
-	static int dodawanie(int a, int b) {
-		int wynik = a + b;
+	static float dodawanie(float a, float b) {
+		float wynik = a + b;
 		return wynik;
 	}
